pyproject.toml
[project]
name = "web-scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.23",
    "langchain-openai>=1.1.12",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tinydb>=4.8.2",
    "torchvision>=0.26.0",
]
